:recycle: Sprite tests per connector
