+++ Assets/Tests/PlayMode/Scripts/AnimationTriggerTest.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3832339a800540e3a4c9bb59224cdf8a
+timeCreated: 1564041591
